+++ tools/system_libs.py
@@ -1378,6 +1378,15 @@ class libcxxabi(NoExceptLibrary, MTLibrary, DebugLibrary):
     ]
   includes = ['system/lib/libcxx/src']
 
+  def __init__(self, **kwargs):
+    super().__init__(**kwargs)
+    # TODO EXCEPTION_STACK_TRACES currently requires the debug version of
+    # libc++abi, causing the debug version of libc++abi to be linked, which
+    # increases code size. libc++abi is not a big library to begin with, but if
+    # this becomes a problem, consider making EXCEPTION_STACK_TRACES work with
+    # the non-debug version of libc++abi.
+    self.is_debug |= settings.EXCEPTION_STACK_TRACES
+
   def get_cflags(self):
     cflags = super().get_cflags()
     if not self.is_mt and not self.is_ww: